requirements.txt
langchain==0.3
python-dotenv==1.1.0
langchain-openai==0.3.19
streamlit==1.32.0